pyproject.toml
[tool.poetry]
name = "install_rate"
version = "0.1.0"
description = "Predicting the probability that a user will install an app after seeing an ad for it"
authors = ["corentin.garet <garetcorentin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.15.1"
gcsfs = "^2024.10.0"
numpy = "^2.1.3"
pandas = "^2.2.3"
db-dtypes = "^1.3.1"
google-cloud-storage = "^2.19.0"
google-cloud-bigquery = "^3.27.0"
google-cloud-bigquery-storage = "^2.27.0"
pyyaml = "^6.0.2"
scikit-learn = "^1.6.0"

[tool.poetry.group.ci.dependencies]
mypy = "^1.13.0"
ruff = "^0.8.2"
vulture = "^2.13"
pandas-stubs = "^2.2.3.241126"
mypy-extensions = "^1.0.0"

[tool.mypy]
ignore_missing_imports = true
